Add dynamic content to header via js
The dynamic content include:
- the name, which is the main h1 title
- the "about me" description
- a portrait image
- social links

The name, about me description, and social links are grabbed from the
info.json file in the customData folder.

Note howerver, that the social link list is not initialized, yet. The
code will be implemented in its own module, for possible reuse in the
footer.

The portrait image must be a jpeg file named "portrait.jpeg", and has to
be placed in the customData folder.

Placeholder data and images are used.