Add dynamic list of social links to .about-me__social-list ul in header via js
Data are grabbed from customData/info.json.

The list items associated to each social link is composed of an anchor
element and an icon.

Animated lottie icons have been used, which animate on mouse hover or
tap. Note that in info.json, social links are specified as an object,
whose keys must coincide with the icon filename.