Adapt style of .about-me element in the header for tablet view
The text in the p element in the .about-me element must be wrapped
around the image in the header. However, such image is a sibling of the
.about-me div. Hence, setting it to float does not affect the text in
such p element.

Therefore, a div with class .about-me__float-space is added inside the
.about-me div, which has the css property "float: left".

The size of the grid rows of the header and of such
.about-me__float-space div have been opportunely set, to make the float
element match the part of the image which is overlapped to the .about-me
div.